# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Episode Api's",
        "description": "Swagger creates human-readable documentation for your APIs.",
        "version": "1.0.0"
    },
    "host": "episodicserver.ignivastaging.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/admin/reviewListing": {
            "get": {
                "tags": [
                    "ADMIN REVIEW APIS"
                ],
                "summary": "review listing",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Enter Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by user",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Failed"
                    },
                    "405": {
                        "description": "Undocumented data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/reviewDetail": {
            "get": {
                "tags": [
                    "ADMIN REVIEW APIS"
                ],
                "summary": "review detail",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Enter Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "review_id",
                        "in": "query",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Failed"
                    },
                    "405": {
                        "description": "Undocumented data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/postReview": {
            "post": {
                "tags": [
                    "ADMIN REVIEW APIS"
                ],
                "summary": "Post review",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Enter Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "review_id",
                        "in": "formData",
                        "description": "review id",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "status(1=>approved,2=>declined)",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "reason",
                        "in": "formData",
                        "description": "if rating declined ,then this filed is required",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Failed"
                    },
                    "405": {
                        "description": "Undocumented data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}